# specs/001-recipe-sharing/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Oppskrift Recipe API
  description: |
    API for creating, organizing, and sharing recipes on the Oppskrift federated platform.
    Supports ActivityPub federation and Schema.org Recipe vocabulary.
  version: 1.0.0
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: https://{instance}/api/v1
    variables:
      instance:
        default: localhost:3000
        description: Oppskrift instance hostname
...
paths:
...
  /recipes/{id}:
    parameters:
      - $ref: '#/components/parameters/RecipeId'

    get:
      summary: Get recipe by ID
      tags: [Recipes]
      responses:
        '200':
          description: Recipe details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
            application/ld+json:
              schema:
                $ref: '#/components/schemas/RecipeJsonLd'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update recipe
      tags: [Recipes]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeUpdate'
      responses:
        '200':
          description: Recipe updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete recipe
      tags: [Recipes]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Recipe deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /books/{id}:
    parameters:
      - $ref: '#/components/parameters/BookId'

    get:
      summary: Get recipe book by ID
      tags: [Recipe Books]
      responses:
        '200':
          description: Recipe book details with recipes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeBookWithRecipes'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update recipe book
      tags: [Recipe Books]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeBookUpdate'
      responses:
        '200':
          description: Recipe book updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeBook'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete recipe book
      tags: [Recipe Books]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Recipe book deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /books/{id}/recipes/{recipeId}:
    parameters:
      - $ref: '#/components/parameters/BookId'
      - name: recipeId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    delete:
      summary: Remove recipe from book
      tags: [Recipe Books]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Recipe removed from book
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ============ SOCIAL ============
...
  /users/{id}:
    parameters:
      - $ref: '#/components/parameters/UserId'

    get:
      summary: Get user profile
      tags: [Users]
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RecipeId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    BookId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
  schemas:
...
    Recipe:
      type: object
      properties:
        id:
          type: string
          format: uuid
        author:
          $ref: '#/components/schemas/UserSummary'
        title:
          type: string
          maxLength: 200
        description:
          type: string
        visibility:
          $ref: '#/components/schemas/Visibility'
        prep_time_min:
          type: integer
          minimum: 0
        cook_time_min:
          type: integer
          minimum: 0
        total_time_min:
          type: integer
          minimum: 0
          readOnly: true
        servings:
          type: string
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/Ingredient'
          maxItems: 50
        instructions:
          type: array
          items:
            $ref: '#/components/schemas/InstructionStep'
          maxItems: 30
        images:
          type: array
          items:
            $ref: '#/components/schemas/RecipeImage'
          maxItems: 10
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        ap_id:
          type: string
          format: uri
      required:
        - id
        - author
        - title
        - visibility
        - ingredients
        - instructions
        - created_at
        - updated_at
...
    RecipeUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
        visibility:
          $ref: '#/components/schemas/Visibility'
        prep_time_min:
          type: integer
          minimum: 0
        cook_time_min:
          type: integer
          minimum: 0
        servings:
          type: string
          maxLength: 50
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        ingredients:
          type: array
          items:
            $ref: '#/components/schemas/IngredientCreate'
          minItems: 1
          maxItems: 50
        instructions:
          type: array
          items:
            $ref: '#/components/schemas/InstructionStepCreate'
          minItems: 1
          maxItems: 30
...
    RecipeSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        author:
          $ref: '#/components/schemas/UserSummary'
        title:
          type: string
        description:
          type: string
        primary_image:
          $ref: '#/components/schemas/RecipeImage'
        total_time_min:
          type: integer
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        created_at:
          type: string
          format: date-time
    RecipeJsonLd:
      type: object
      description: Schema.org Recipe in JSON-LD format
      properties:
        '@context':
          type: string
          enum: ['https://schema.org']
        '@type':
          type: string
          enum: ['Recipe']
        name:
          type: string
        description:
          type: string
        author:
          type: object
          properties:
            '@type':
              type: string
              enum: ['Person']
            name:
              type: string
        prepTime:
          type: string
          description: ISO 8601 duration
        cookTime:
          type: string
        totalTime:
          type: string
        recipeYield:
          type: string
        recipeIngredient:
          type: array
          items:
            type: string
        recipeInstructions:
          type: array
          items:
            type: object
            properties:
              '@type':
                type: string
                enum: ['HowToStep']
              text:
                type: string
        image:
          type: array
          items:
            type: string
            format: uri
        datePublished:
          type: string
          format: date-time
        dateModified:
          type: string
          format: date-time
...
    Ingredient:
      type: object
      properties:
        id:
          type: string
          format: uuid
        position:
          type: integer
          minimum: 1
          maximum: 50
        quantity:
          type: number
          description: Amount in metric units
        quantity_display:
          type: string
          description: Formatted for user's measurement preference
        unit:
          type: string
        name:
          type: string
        notes:
          type: string
    IngredientCreate:
      type: object
      properties:
        quantity:
          type: number
        unit:
          type: string
          maxLength: 20
        name:
          type: string
          minLength: 1
          maxLength: 200
        notes:
          type: string
          maxLength: 200
      required:
        - name
    InstructionStep:
      type: object
      properties:
        id:
          type: string
          format: uuid
        step_number:
          type: integer
          minimum: 1
          maximum: 30
        description:
          type: string
        image_url:
          type: string
          format: uri
        duration_min:
          type: integer
          minimum: 0
    InstructionStepCreate:
      type: object
      properties:
        description:
          type: string
          minLength: 1
          maxLength: 2000
        duration_min:
          type: integer
          minimum: 0
      required:
        - description
    RecipeImage:
      type: object
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        alt_text:
          type: string
        position:
          type: integer
        is_primary:
          type: boolean

    # ============ RECIPE BOOK SCHEMAS ============
    RecipeBook:
      type: object
      properties:
        id:
          type: string
          format: uuid
        owner:
          $ref: '#/components/schemas/UserSummary'
        title:
          type: string
        description:
          type: string
        cover_image_url:
          type: string
          format: uri
        visibility:
          $ref: '#/components/schemas/Visibility'
        recipe_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RecipeBookWithRecipes:
      allOf:
        - $ref: '#/components/schemas/RecipeBook'
        - type: object
          properties:
            recipes:
              type: array
              items:
                $ref: '#/components/schemas/RecipeSummary'
...
    RecipeBookUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
        visibility:
          $ref: '#/components/schemas/Visibility'
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        display_name:
          type: string
        bio:
          type: string
        avatar_url:
          type: string
          format: uri
        measurement_pref:
          $ref: '#/components/schemas/MeasurementPref'
        created_at:
          type: string
          format: date-time
        ap_id:
          type: string
          format: uri
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        display_name:
          type: string
        avatar_url:
          type: string
          format: uri
    UserProfile:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            recipe_count:
              type: integer
            book_count:
              type: integer
            follower_count:
              type: integer
            following_count:
              type: integer
            is_following:
              type: boolean
              description: Whether current user follows this user
...
    Visibility:
      type: string
      enum: [public, private]
      default: public
    Difficulty:
      type: string
      enum: [easy, medium, hard]
    MeasurementPref:
      type: string
      enum: [metric, imperial]
      default: metric
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    ValidationError:
      type: object
      properties:
        error:
          type: string
          enum: [validation_error]
        message:
          type: string
        fields:
          type: object
          additionalProperties:
            type: array
            items:
              type: string